+++ test/neocaml-navigation-test.el
@@ -408,6 +408,14 @@ end"
       (expect (looking-at "struct") :to-be-truthy)
       (expect (line-number-at-pos) :to-equal 1)))
 
+  (it "jumps out of an optional parameter with default"
+    (with-neocaml-buffer "let bar ?(foo = 123) = assert false"
+      (goto-char (point-min))
+      (search-forward "123")
+      (backward-char 2)
+      (neocaml-backward-up-list)
+      (expect (char-after) :to-equal ??)))
+
   (it "signals an error at the top level"
     (with-neocaml-buffer "let x = 1"
       (goto-char (point-min))